Save labels firehose cursor from queue instead of from latest message
When the server reloads/shuts down any labels waiting to be processed
are discarded. By saving the `seq` value of the first item in queue to
database instead of every 10th labels message, we can now use it at
next startup to recover any labels in queue that were lost.

Cursor update behavior in the client object stays the same as it's
reused between connections while the server is running.