Make errors caused by trying to add existing post into DB quieter
We already know that trying to add post info with the same CID will
fail with an `IntegrityError` so replace tracebacks with either a
simple log entry (if logger level is set to debug) or none at all.
Useful when replaying firehose events either due to a reconnect or
in experiments.

Logs with tracebacks are still outputted if the error is caused by
something else

Also improve the added post counter to not count posts already in DB